fix danger lint warning
