+++ lib/checkstyle_format/gem_version.rb
@@ -1,3 +1,3 @@
 module CheckstyleFormat
-  VERSION = "0.0.3".freeze
+  VERSION = "0.0.4".freeze
 end
