+++ lib/checkstyle_format/plugin.rb
@@ -1,3 +1,5 @@
+require_relative "checkstyle_error"
+
 module Danger
   # Danger plugin for checkstyle formatted xml file.
   #
@@ -27,19 +29,6 @@ module Danger
         send_inline_comment(errors)
       else
         raise "not implemented." # TODO: not implemented.
-      end
-    end
-
-    CheckstyleError = Struct.new(:file_name, :line, :column, :severity, :message, :source) do
-      def self.generate(node, parent_node, base_path)
-        CheckstyleError.new(
-          parent_node[:name].sub(/^#{base_path}/, ""),
-          node[:line].to_i,
-          node[:column].nil? ? nil : node[:column].to_i,
-          node[:severity],
-          node[:message],
-          node[:source]
-        )
       end
     end
